Do not log context canceled error in Close. Move error APIs to utils/errors.go
